requirements.txt
backports.csv==1.0.7
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-allauth==0.40.0
django-environ==0.4.5
django-import-export==1.2.0
environ==1.0
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.0
Pillow==6.2.0
psycopg2==2.8.3
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.6
xlrd==1.2.0
xlwt==1.3.0
